New code from lwip project
lwip repo:
f69b184 Fix compile when netconn api is disabled but socket api is enabled. Problem is that declaring functions as static in a public header will produce warnings in every file it is included because the static functions are not implemented. Solution: When socket api is enabled, netconn is simply available, too (Socket api uses netconn api internally)

lwip-contrib repo:
642f3b1 Fix compile without LWIP_TCP or LWIP_NETCONN